requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.16.0
chess==1.10.0
cssselect2==0.7.0
defusedxml==0.7.1
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.4.0
pycairo==1.26.1
pycparser==2.22
python-dotenv==1.0.1
tinycss2==1.3.0
webencodings==0.5.1
yarl==1.9.4
